Disable Module verifier, rename Settings to MullvadSettings to avoid clash with Apple private Framework of the same name
