+++ ios/MullvadSettings/DeviceState.swift
