+++ ios/MullvadSettings/SettingsManager.swift
