+++ ios/MullvadSettings/SettingsStore.swift
