+++ ios/MullvadSettings/StoredDeviceData.swift
