+++ ios/MullvadVPN/AppDelegate.swift
@@ -9,11 +9,11 @@
 import BackgroundTasks
 import MullvadLogging
 import MullvadREST
+import MullvadSettings
 import MullvadTransport
 import MullvadTypes
 import Operations
 import RelayCache
-import Settings
 import StoreKit
 import UIKit
 import UserNotifications
