+++ ios/MullvadVPN/SimulatorTunnelProvider/SimulatorTunnelProviderHost.swift
@@ -11,12 +11,12 @@
 import Foundation
 import MullvadLogging
 import MullvadREST
+import MullvadSettings
 import MullvadTransport
 import MullvadTypes
 import enum NetworkExtension.NEProviderStopReason
 import RelayCache
 import RelaySelector
-import Settings
 import TunnelProviderMessaging
 
 final class SimulatorTunnelProviderHost: SimulatorTunnelProviderDelegate {
