+++ ios/MullvadVPN/TunnelManager/LoadTunnelConfigurationOperation.swift
@@ -8,9 +8,9 @@
 
 import Foundation
 import MullvadLogging
+import MullvadSettings
 import MullvadTypes
 import Operations
-import Settings
 
 class LoadTunnelConfigurationOperation: ResultOperation<Void> {
     private let logger = Logger(label: "LoadTunnelConfigurationOperation")
