+++ ios/MullvadVPN/TunnelManager/UpdateAccountDataOperation.swift
@@ -9,9 +9,9 @@
 import Foundation
 import MullvadLogging
 import MullvadREST
+import MullvadSettings
 import MullvadTypes
 import Operations
-import Settings
 
 class UpdateAccountDataOperation: ResultOperation<Void> {
     private let logger = Logger(label: "UpdateAccountDataOperation")
