+++ ios/MullvadVPN/View controllers/Account/AccountInteractor.swift
@@ -8,9 +8,9 @@
 
 import Foundation
 import MullvadREST
+import MullvadSettings
 import MullvadTypes
 import Operations
-import Settings
 import StoreKit
 
 final class AccountInteractor {
