+++ ios/MullvadVPN/View controllers/OutOfTime/OutOfTimeInteractor.swift
@@ -9,9 +9,9 @@
 import Foundation
 import MullvadLogging
 import MullvadREST
+import MullvadSettings
 import MullvadTypes
 import Operations
-import Settings
 import StoreKit
 
 final class OutOfTimeInteractor {
