+++ ios/MullvadVPN/View controllers/Settings/SettingsDataSource.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2021 Mullvad VPN AB. All rights reserved.
 //
 
-import Settings
+import MullvadSettings
 import UIKit
 
 final class SettingsDataSource: UITableViewDiffableDataSource<
