+++ ios/PacketTunnel/DeviceCheck/DeviceStateAccessorProtocol.swift
@@ -7,7 +7,7 @@
 //
 
 import Foundation
-import Settings
+import MullvadSettings
 
 /// A protocol that formalizes device state accessor dependency used by `DeviceCheckOperation`.
 protocol DeviceStateAccessorProtocol {
